Add an example s7check.sd7 program
This is a quick draft taken from an example provided by Thomas Mertes
here: https://github.com/ThomasMertes/seed7/issues/34#issuecomment-2788565294

I just modified the output, removing the leading "*** " from the output
message, to make this compatible with Emacs compilation-mode.

More work would be needed to make this a full-blown checker program: checking
if a file name is provided, is valid, help, etc...  But its working correctly
as it is as long as a valid Seed7 file is provided as the first argument on
the command line.