PostgreSQL - Money Type Support
Adds select (i.e. column) and bind support for the money type in PostgreSQL.

Internally, Postgres seems to store money as an integer representing the lowest monetary unit for the database's locale. Upon receiving such values, Postgres automatically converts the input value to the locale's standard unit (such as dollars); but selection returns the un-converted value (such a cents). Thus, I include locale.h, load the database's current locale setting for money, then use `frac_digits` to form a monetary denominator for which to automate the conversion when fetching values so that the user/developer can use the same values for both input and output.

I didn't achieve a full understanding of the string to number conversions that are employed within `sql_post.c`, but rather copied and adjusted code for similar values (integers and floats/numerics), and tested the input/output against a PostgreSQL database. I say this, to indicate it might be worth double-checking those conversions--even though it seems to work properly.