Basic HTTP/S Post Functionality
Built off of the existing `gethttp` and `gethttps` libraries. Allows the dev to post any type of content via the new `httpBody` structure; though, I made functions for posting `text/plain`, `application/json` and single-tier `application/x-www-form-urlencoded` (that is, I didn't implement array values) content.

It would be nice, at some point, to have a response code for the get/post functions, as some error codes for APIs are used to indicate a request change that could be automated.